[♻️refactor]: 스타일 컨밴션에 맞게 S dot 네이밍 및 .styles 파일명으로 변경
