+++ src/components/FlipClock/Digit/Digit.styles.tsx
