+++ src/components/FlipClock/FlipClock.styles.tsx
